Доработанный автотест исправлены ошибки
